git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@5031 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/cl_screen.c
@@ -1462,7 +1462,7 @@ void SCR_DrawFPS (void)
 	double t;
 	extern int fps_count;
 	static float lastfps;
-	static float deviationtimes[64];
+	static double deviationtimes[64];
 	static int deviationframe;
 	char str[80];
 	int sfps, frame;
